requirements.txt
 streamlit==1.41.1
 httpx==0.27.2
 ontology-verbalizer==1.1.2
 openai==1.12.0
 pandas==2.2.1